add type hints and renamed the field / module
