+++ custom_addons/manufacturable-field/views/product_views.xml
@@ -7,7 +7,7 @@
         <field name="inherit_id" ref="product.product_template_only_form_view"/>
         <field name="arch" type="xml">
             <xpath expr="//field[@name='barcode']" position="after">
-                <field name="can_be_manufactured"/>
+                <field name="manufacturable"/>
             </xpath>
         </field>
     </record>
@@ -19,7 +19,7 @@
         <field name="inherit_id" ref="product.product_normal_form_view"/>
         <field name="arch" type="xml">
             <xpath expr="//field[@name='barcode']" position="after">
-                <field name="can_be_manufactured"/>
+                <field name="manufacturable"/>
             </xpath>
         </field>
     </record>
@@ -31,7 +31,7 @@
         <field name="inherit_id" ref="product.product_template_tree_view"/>
         <field name="arch" type="xml">
             <xpath expr="//field[@name='qty_available']" position="after">
-                <field name="can_be_manufactured" optional="show" readonly="1"/>
+                <field name="manufacturable" optional="show" readonly="1"/>
             </xpath>
         </field>
     </record>
@@ -43,7 +43,7 @@
         <field name="inherit_id" ref="product.product_product_tree_view"/>
         <field name="arch" type="xml">
             <xpath expr="//field[@name='qty_available']" position="after">
-                <field name="can_be_manufactured" optional="show" readonly="1"/>
+                <field name="manufacturable" optional="show" readonly="1"/>
             </xpath>
         </field>
     </record>
